+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### vNEXT
 
+### v3.0.2
+
 - **action**: `schema` can now point to GraphQL API
 - **cli**: fix missing `--onUsage` in the diff command